pyproject.toml
[project]
name = "code-review-agent"
version = "0.1.0"
description = "AI-powered code review agent integrated with Azure DevOps"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.2.0",
    "langgraph>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "azure-devops>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "gitpython>=3.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]
